requirements.txt
# Core dependencies
langchain>=0.0.267
langchain-core>=0.0.15
langchain-huggingface>=0.0.1
langchain-community>=0.0.1
llama-index>=0.9.0
llama-index-core>=0.10.0
llama-index-vector-stores-faiss>=0.1.0
llama-index-embeddings-huggingface>=0.1.0

# Machine learning and NLP
transformers>=4.35.0
peft>=0.5.0
torch>=2.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
bitsandbytes>=0.41.0
accelerate>=0.21.0

# Document processing
PyPDF2>=3.0.0
pypdf>=3.15.0

# Web UI
streamlit>=1.26.0
streamlit-modal>=0.1.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.5
pydantic>=2.0.0
huggingface-hub>=0.16.4

# Optional: For GPU support
# torch>=2.0.0+cu118  # Uncomment and specify correct CUDA version if using GPU
